requirements.txt
aiodns==3.0.0
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
apturl==0.5.2
async-timeout==4.0.2
attrs==23.1.0
bcrypt==3.2.0
blinker==1.6.2
Brlapi==0.8.3
certifi==2020.6.20
cffi==1.15.1
chardet==4.0.0
charset-normalizer==3.1.0
cliapp==1.20180812.1
click==8.1.3
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
cssselect==1.2.0
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
distro==1.7.0
distro-info===1.1build1
dnspython==2.3.0
duplicity==0.8.21
fastapi==0.95.2
fasteners==0.14.1
filelock==3.12.0
Flask==2.3.2
frozenlist==1.3.3
future==0.18.2
gyp==0.1
h11==0.14.0
html-similarity==0.3.3
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
itsdangerous==2.1.2
jeepney==0.7.1
Jinja2==3.1.2
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
loguru==0.7.0
louis==3.20.0
lxml==4.9.2
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.3.6
MarkupSafe==2.1.2
monotonic==1.6
more-itertools==8.10.0
multidict==6.0.4
nest-asyncio==1.5.6
netifaces==0.11.0
oauthlib==3.2.0
olefile==0.46
paramiko==2.9.3
parsel==1.6.0
pexpect==4.8.0
Pillow==9.0.1
protobuf==3.12.4
ptyprocess==0.7.0
pycairo==1.20.1
pycares==4.3.0
pycparser==2.21
pycups==2.0.1
pydantic==1.10.7
Pygments==2.11.2
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
python-apt==2.4.0+ubuntu1
python-dateutil==2.8.1
python-debian===0.1.43ubuntu1
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.25.1
requests-file==1.5.1
SecretStorage==3.3.1
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
systemd-python==234
tldextract==3.4.1
ttystatus==0.38
typing_extensions==4.5.0
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
usb-creator==0.3.7
uvicorn==0.22.0
w3lib==2.1.1
wadllib==1.3.6
Werkzeug==2.3.4
whois==0.9.27
xdg==5
xkit==0.0.0
yarl==1.9.2
zipp==1.0.0
